Move tests to its module and fix function
